+++ modules/src/image/detection/ll_boundaries.c
@@ -57,7 +57,7 @@ static float min_contrast(Flist l, float *length)
 {
   double per;
   float mu,minmu,x,y,ox,oy;
-  int i,ix,iy;
+  int i, ix, iy;
 
   per = 0.;
   minmu = FLT_MAX;
@@ -72,8 +72,8 @@ static float min_contrast(Flist l, float *length)
       ox = x; oy = y;
     }
     
-    ix = (int) floor((double) x + .5) - 1;
-    iy = (int) floor((double) y + .5) - 1;
+    ix = floor((double) x + .5) - 1;
+    iy = floor((double) y + .5) - 1;
     if (ix>=0 && iy>=0 && ix<NormOfDu->ncol && iy<NormOfDu->nrow) {
       mu = NormOfDu->gray[NormOfDu->ncol*iy+ix];
       if (mu<minmu) minmu=mu;
